pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.4.1",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-experimental>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-tavily>=0.2.17",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "tavily-python>=0.7.23",
    "uvicorn>=0.41.0",
    "wikipedia>=1.4.0",
    "youtube-search>=2.2.0",
    "youtube-search-python>=1.6.6",
]
